Added skeleton for accel and gyro odr
Noticed that current definition methods for member variables is quite ineficient, with repeated names and functionality. Probably, the accel and gyro data could be neatly packed in structs.

Moreover, the configurable values follow the PURPOSE::SENSOR definition, when it probably makes more sense to do it the other way around. Fix